+++ src/java/lazytest/ExpectationFailed.java
@@ -2,15 +2,23 @@ package lazytest;
 
 import clojure.lang.IExceptionInfo;
 import clojure.lang.IPersistentMap;
+import clojure.lang.PersistentHashMap;
 
-public class ExpectationFailed extends Throwable implements IExceptionInfo {
+public class ExpectationFailed extends AssertionError implements IExceptionInfo {
     public final IPersistentMap reason;
 
+    private static final long serialVersionUID = 1L;
+    private static final String defaultMessage = "Expectation failed";
+
     public ExpectationFailed(IPersistentMap reason) {
-        super("Expectation Failed");
+        this(defaultMessage, reason);
+    }
+
+    public ExpectationFailed(String message, IPersistentMap reason) {
+        super(message == null ? defaultMessage : message);
 
         if (reason == null) {
-            throw new IllegalArgumentException("Additional data must be non-nil.");
+            reason = PersistentHashMap.EMPTY;
         }
         this.reason = reason;
     }